item Type

+++ app/src/main/java/com/sideral/ovejas/model/GridItem.kt
@@ -1,7 +1,7 @@
 package com.sideral.ovejas.model
 
 class GridItem(title: String) : Item(title) {
-    override fun getItemType(): Short {
+    override fun getItemType(): Int {
         return GRID_ITEM_TYPE;
     }
